+++ src/backend/ci/ext/tencent/process/biz-process-tencent/src/main/kotlin/com/tencent/devops/process/engine/atom/task/SpmDistributionTaskAtom.kt
@@ -271,7 +271,7 @@ class SpmDistributionTaskAtom @Autowired constructor(
                 logger.error("Response failed. msg: ${responseJson["msg"].asString}")
                 throw BuildTaskException(
                     errorType = ErrorType.SYSTEM,
-                    errorCode = ERROR_BUILD_TASK_CDN_FAIL,
+                    errorCode = ERROR_BUILD_TASK_CDN_FAIL.toInt(),
                     errorMsg = "分发CDN失败"
                 )
             }
